CLI watch flag for get commands, timing fixes, tenant in stats, UX polish
