+++ agent/api/v1/model.go
@@ -145,6 +145,7 @@ type ExportDetailListResponse struct {
 }
 
 type StatsResponse struct {
+	TenantName string               `json:"tenant_name"`
 	Statfs StatfsResponse          `json:"statfs"`
 	Btrfs  FilesystemStatsResponse `json:"btrfs"`
 }
