+++ ctl/ctl.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"time"
 
-	"github.com/erikmagkekse/btrfs-nfs-csi/utils"
 	"github.com/urfave/cli/v3"
 )
 
@@ -15,89 +14,21 @@ var (
 	Commit  = "unknown"
 )
 
-func showStats(ctx context.Context, cmd *cli.Command) error {
-	return runWatch(ctx, cmd, func() error {
-		resp, err := apiClient.Stats(ctx)
-		if err != nil {
-			return err
-		}
-		return output(cmd, resp, func() {
-			fmt.Println("statfs:")
-			fmt.Printf("  Total:       %s\n", utils.FormatBytes(resp.Statfs.TotalBytes))
-			fmt.Printf("  Used:        %s (%.0f%%)\n", utils.FormatBytes(resp.Statfs.UsedBytes), usedPct(resp.Statfs.UsedBytes, resp.Statfs.TotalBytes))
-			fmt.Printf("  Free:        %s\n", utils.FormatBytes(resp.Statfs.FreeBytes))
-			fmt.Println()
-			fmt.Println("btrfs:")
-			fmt.Printf("  Total:       %s\n", utils.FormatBytes(resp.Btrfs.TotalBytes))
-			fmt.Printf("  Used:        %s (%.0f%%)\n", utils.FormatBytes(resp.Btrfs.UsedBytes), usedPct(resp.Btrfs.UsedBytes, resp.Btrfs.TotalBytes))
-			fmt.Printf("  Free:        %s\n", utils.FormatBytes(resp.Btrfs.FreeBytes))
-			fmt.Printf("  Unallocated: %s\n", utils.FormatBytes(resp.Btrfs.UnallocatedBytes))
-			fmt.Printf("  Metadata:    %s / %s\n", utils.FormatBytes(resp.Btrfs.MetadataUsedBytes), utils.FormatBytes(resp.Btrfs.MetadataTotalBytes))
-			fmt.Printf("  Data ratio:  %.1f\n", resp.Btrfs.DataRatio)
-			fmt.Println()
-			fmt.Println("devices:")
-			w := tab()
-			if isWide(cmd) {
-				_, _ = fmt.Fprintln(w, "  DEVICE\tSIZE\tALLOCATED\tREAD\tWRITTEN\tREAD_IOS\tWRITE_IOS\tREAD_ERR\tWRITE_ERR\tFLUSH_ERR\tCSUM_ERR\tGEN_ERR\tSTATUS")
-				for _, d := range resp.Btrfs.Devices {
-					status := "ok"
-					if d.Missing {
-						status = "MISSING"
-					} else if d.Errors.ReadErrs+d.Errors.WriteErrs+d.Errors.FlushErrs+d.Errors.CorruptionErrs+d.Errors.GenerationErrs > 0 {
-						status = "ERRORS"
-					}
-					_, _ = fmt.Fprintf(w, "  %s\t%s\t%s\t%s\t%s\t%d\t%d\t%d\t%d\t%d\t%d\t%d\t%s\n",
-						d.Device, utils.FormatBytes(d.SizeBytes), utils.FormatBytes(d.AllocatedBytes),
-						utils.FormatBytes(d.IO.ReadBytesTotal), utils.FormatBytes(d.IO.WriteBytesTotal),
-						d.IO.ReadIOsTotal, d.IO.WriteIOsTotal,
-						d.Errors.ReadErrs, d.Errors.WriteErrs, d.Errors.FlushErrs,
-						d.Errors.CorruptionErrs, d.Errors.GenerationErrs, status)
-				}
-			} else {
-				_, _ = fmt.Fprintln(w, "  DEVICE\tSIZE\tALLOCATED\tREAD\tWRITTEN\tERRORS\tSTATUS")
-				for _, d := range resp.Btrfs.Devices {
-					errs := d.Errors.ReadErrs + d.Errors.WriteErrs + d.Errors.FlushErrs + d.Errors.CorruptionErrs + d.Errors.GenerationErrs
-					status := "ok"
-					if d.Missing {
-						status = "MISSING"
-					} else if errs > 0 {
-						status = "ERRORS"
-					}
-					_, _ = fmt.Fprintf(w, "  %s\t%s\t%s\t%s\t%s\t%d\t%s\n",
-						d.Device, utils.FormatBytes(d.SizeBytes), utils.FormatBytes(d.AllocatedBytes),
-						utils.FormatBytes(d.IO.ReadBytesTotal), utils.FormatBytes(d.IO.WriteBytesTotal),
-						errs, status)
-				}
-			}
-			_ = w.Flush()
-		})
-	})
-}
-
-func showHealth(ctx context.Context, cmd *cli.Command) error {
-	resp, err := apiClient.Healthz(ctx)
-	if err != nil {
-		return err
-	}
-	return output(cmd, resp, func() {
-		fmt.Printf("status:  %s\nversion: %s\ncommit:  %s\nuptime:  %ds\n",
-			resp.Status, resp.Version, resp.Commit, resp.UptimeSeconds)
-	})
-}
-
 func Run(args []string) {
 	app := &cli.Command{
 		Name:  "btrfs-nfs-csi",
 		Usage: "btrfs-nfs-csi agent CLI",
+		OnUsageError: func(_ context.Context, _ *cli.Command, err error, _ bool) error {
+			return err
+		},
 		Before: func(ctx context.Context, cmd *cli.Command) (context.Context, error) {
 			initClient(cmd)
 			cmdStart = time.Now()
 			return ctx, nil
 		},
 		After: func(ctx context.Context, cmd *cli.Command) error {
-			if !isJSON(cmd) && !cmd.IsSet("watch") && cmd.String("columns") == "" {
-				printTiming()
-				fmt.Println()
+			if !cmdStart.IsZero() && !ranWatch && !isJSON(cmd) && cmd.String("columns") == "" {
+				timingLine = fmtTiming(time.Since(cmdStart))
 			}
 			return nil
 		},
@@ -111,35 +42,17 @@ func Run(args []string) {
 			snapshotCmd(),
 			exportCmd(),
 			taskCmd(),
-			{
-				Name:  "version",
-				Usage: "show CLI version",
-				Action: func(ctx context.Context, cmd *cli.Command) error {
-					fmt.Printf("btrfs-nfs-csi %s (%s)\n", Version, Commit)
-					if cmd.Root().String("agent-url") != "" && cmd.Root().String("agent-token") != "" {
-						if h, err := apiClient.Healthz(ctx); err == nil {
-							fmt.Printf("agent         %s (%s)\n", h.Version, h.Commit)
-						}
-					}
-					return nil
-				},
-			},
-			{
-				Name:   "stats",
-				Usage:  "show filesystem stats",
-				Flags:  []cli.Flag{watchFlag()},
-				Action: showStats,
-			},
-			{
-				Name:   "health",
-				Usage:  "show agent health",
-				Action: showHealth,
-			},
+			versionCmd(),
+			statsCmd(),
+			healthCmd(),
 		},
 	}
 
 	if err := app.Run(context.Background(), append([]string{"btrfs-nfs-csi"}, injectWatchDefault(args)...)); err != nil {
 		_, _ = fmt.Fprintf(os.Stderr, "error: %v\n", err)
 		os.Exit(1)
+	}
+	if timingLine != "" {
+		fmt.Fprintf(os.Stderr, "\n%s\n", timingLine)
 	}
 }
